Add files via upload
Use the code provided in demo 02 to complete this assignment.

Step 1. Follow along with the tutorial to gain an understanding of the process

Step 2. In a new .ipynb notebook, reproduce the results utilizing the original MNIST dataset (built into tf.keras.datasets)

Step 3. Report on the results in terms of prediction accuracy on the train and test datasets

Step 4. Choose one of the proposed modifications below:

Add another Dense layer of 128 nodes
Increase the current number of nodes in the layer to 256
Hypothesize how it would change the performance results

Step 5. Modify the model based on the chosen method and train

Step 6. Report on the results of the modified model and if it matches your hypothesis

Step 7. Experiment with different optimizers, loss functions, dropout, regularizations, activation functions and observe the change in performance as you tune these hyper parameters.